launcher: dependency inject container image into ContainerRunner
Adding dependency injection to ContainerRunner will allow it to be
tested using mocks/fakes when real resources are unavailable in a test
environment.

Previously, the runner pulled the container image internally during initialization.
This change moves the image-pulling responsibility up to the launcher orchestration layer.
